consolidating code
