+++ Common/Data/DatabaseAdapter.cs
@@ -1,4 +1,5 @@
 ﻿using System.Configuration;
+using System.Data;
 
 namespace Betlln.Data
 {
@@ -22,5 +23,20 @@ namespace Betlln.Data
         }
 
         protected string ConnectionAddress { get; set; }
+        
+        protected static T? ReadNullableValue<T>(IDataReader reader, string columnName)
+            where T : struct 
+        {
+            return (T?) (reader.IsDBNull(reader.GetOrdinal(columnName))
+                ? null
+                : reader[columnName]);
+        }
+
+        protected static string ReadString(IDataReader reader, string columnName)
+        {
+            return reader.IsDBNull(reader.GetOrdinal(columnName))
+                ? null
+                : reader[columnName].ToString();
+        }
     }
 }
